Release 1.02

+++ point/point.h
@@ -9,13 +9,15 @@
 #ifndef point_h
 #define point_h
 
+#include "../env/env.h"
+
 typedef struct point_t {
-    int x;
-    int y;
+    uint8_t x;
+    uint8_t y;
 } point_t;
 
 typedef struct point_namespace {
-    point_t* (*const construct)(int x, int y);
+    point_t* (*const construct)(uint8_t x, uint8_t y);
     int (*const destruct)(point_t* point);
     int (*const distance)(point_t* p1, point_t* p2);
     void (*const print)(point_t* p);
